backend/requirements.txt
fastapi==0.115.14
uvicorn[standard]
sqlalchemy~=2.0.41


aiosqlite~=0.19.0
psycopg2-binary~=2.9.9
aiomysql


pydantic[email]
python-jose[cryptography]
passlib[bcrypt]~=1.7.4
bcrypt~=4.0.1
python-dotenv~=1.1.0
email_validator~=2.2.0
python-multipart~=0.0.20
google-adk~=1.16.0
Authlib~=1.6.0
itsdangerous~=2.2.0
types-SQLAlchemy~=1.4.53.38
requests~=2.32.3
deprecated~=1.2.18
asyncio~=3.4.3
dataclasses~=0.6
httpx

#uv~=0.7.12
apscheduler~=3.11.0
pillow